requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
python-dotenv==1.0.1

